requirements.txt
# Core Dependencies
anthropic>=0.9.0           # Claude API integration
playwright>=1.40.0         # Browser automation
pydantic>=2.5.2           # Data validation
python-dotenv==1.0.0      # Environment management
pytz==2024.1              # Timezone support
aiosqlite==0.19.0         # Async SQLite support
psutil>=5.9.8             # System monitoring (Updated for better metrics)
cachetools==5.3.2         # Caching utilities
httpx>=0.27.0             # HTTP client with proxy support

# Vision & ML
torch==2.2.0              # Deep learning support
torchvision==0.17.0       # Image processing
transformers==4.37.0      # Model implementations
pillow==10.2.0            # Image handling
opencv-python==4.9.0.80   # Computer vision operations

# Database & Caching
redis>=5.0.1              # Fast caching (optional)
sqlalchemy>=2.0.25        # Database ORM
aioredis>=2.0.1          # Async Redis support

# Monitoring & Logging
prometheus-client>=0.19.0  # Metrics collection
opentelemetry-api>=1.21.0 # Distributed tracing
structlog>=24.1.0         # Structured logging
python-json-logger>=2.0.7 # JSON logging support

# Testing & Development
pytest>=8.0.0             # Testing framework
pytest-asyncio>=0.23.5    # Async test support
pytest-cov>=4.1.0         # Coverage reporting
pytest-mock>=3.12.0       # Mocking support
pytest-timeout>=2.2.0     # Test timeouts
pytest-xdist>=3.5.0       # Parallel testing

# Code Quality
black>=24.1.1             # Code formatting
mypy>=1.8.0               # Type checking
ruff>=0.2.1               # Fast Python linter
pre-commit>=3.6.0         # Git hooks 